Moved Closure and Environment classes into messageformat2_context.h
