+++ icu4c/source/i18n/unicode/messageformat2.h
@@ -39,77 +39,6 @@ U_NAMESPACE_BEGIN namespace message2 {
  * @deprecated This API is for technology preview only.
  */
 
-// Intermediate classes used internally in the formatter
-// TODO: these should be private to MessageFormatter, but are used by the ExpressionContext so they aren't
-class Environment;
-// A closure represents the right-hand side of a variable
-// declaration, along with an environment giving values
-// to its free variables
-class Closure : public UMemory {
-    using Expression = MessageFormatDataModel::Expression;
-public:
-    static Closure* create(const Expression&, const Environment&, UErrorCode&);
-    const Expression& getExpr() const {
-        return expr;
-    }
-    const Environment& getEnv() const {
-        return env;
-    }
-    virtual ~Closure();
-private:
-    Closure(const Expression& expression, const Environment& environment) : expr(expression), env(environment) {}
-
-    // An unevaluated expression
-    const Expression& expr;
-    // The environment mapping names used in this
-    // expression to other expressions
-    const Environment& env;
-};
-
-// An environment is represented as a linked chain of
-// non-empty environments, terminating at an empty environment.
-// It's searched using linear search.
-class Environment : public UMemory {
-public:
-    virtual const Closure* lookup(const VariableName&) const = 0;
-    static Environment* create(UErrorCode&);
-    static Environment* create(const VariableName&, Closure*, const Environment&, UErrorCode&);
-    virtual ~Environment();
-};
-
-class NonEmptyEnvironment;
-class EmptyEnvironment : public Environment {
-private:
-    friend class Environment;
-
-    const Closure* lookup(const VariableName&) const;
-    static EmptyEnvironment* create(UErrorCode&);
-    virtual ~EmptyEnvironment();
-    // Adopts its closure argument
-    static NonEmptyEnvironment* create(const VariableName&, Closure*, const Environment&, UErrorCode&);
-
-    EmptyEnvironment() {}
-};
-
-class NonEmptyEnvironment : public Environment {
-private:
-    friend class Environment;
-    const Closure* lookup(const VariableName&) const;
-    // Adopts its closure argument
-    static NonEmptyEnvironment* create(const VariableName&, Closure*, const Environment&, UErrorCode&);
-    virtual ~NonEmptyEnvironment();
-private:
-    friend class Environment;
-
-    NonEmptyEnvironment(const VariableName& v, Closure* c, const Environment& e) : var(v), rhs(c), parent(e) {}
-
-    // Maps VariableName onto Closure*
-    // Chain of linked environments
-    VariableName var;
-    const LocalPointer<Closure> rhs; // should be valid
-    const Environment& parent;
-};
-
 // Note: This class does not currently inherit from the existing
 // `Format` class.
 class U_I18N_API MessageFormatter : UMemory {
